Postpone calculation of ns edits to speed up completion
Sometimes completion knows that for a completion item to be valid the ns
form will have to be edited to require a new namespace. It calculates
the edit for each completion item and includes that in its response.
When there are many of these completion items, the edit calculations
slow down completion.

Now these calculations are postponed to completionItem/resolve, so that
only one edit is calculated, and only after the user selects its item.
See #1053 for background.

This also stops asking clients to resolve documentation if they don't
support that. And it preserves "snippet-kind", which is a secret
backdoor to slip additional data to some clients.

Fixes #1068