Fix unit tests on MS-Windows
Most of the unit test updates are about using the
`test-helper/file-path` and `test-helper/file-uri` function around
test paths where it is missing.

Other important changes
- Introduction of `test-helper/string=` function to use in text
comparisons that treats all Line Endings (LE) as equal. Line endings
can be different on unix (`\n`) vs windows (`\n\r`) which might mess
up with the testing.
- Enhance test-helper/assert-submaps so that it accepts a `set` as
input to disregard order of items. It turns out some lsp vector
results are coming out of maps, and thus their order is
unpredictable. See `handler-test/document-symbol` for an example.
- Use the Apache Commons IO as a new dependency to deal with path
separators in a uniformly way across the differnet architectures where
needed. Most notable use is in
`shared/namespace+source-path->filename`. Note that this brings a new
library in but uses just a small portion of it.
- `cljfmt/reformat-string` is made aware of line endings.
- The default case for drive letter across tests and in
shared/conform-uri is now uppercase. This is to align with how java
always treat the drive letter in the `.toURI` calls. Thus we
uniformaly handle this in the codebase and test as such. Not sure what
external impact would changing the default case have to external
clients.
- Fixes issue with the custom `file:*.jar:subpath` syntax in
`shared/uri->filename`, since that pattern is not valid on windows.
- Update `ci.yml` to unit test on windows but only on JDK-8, and add
macos as an added bonus.